requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3
sniffio==1.3.0
soupsieve==2.4.1
yarl==1.9.2
